message to push

+++ resources/views/admin/events/index.blade.php
@@ -8,11 +8,11 @@
         <div class="col-md-9">
             <div class="card">
                 <div class="card-header">Events</div>
-{{--                <div class="card-body">--}}
-{{--                    <a href="{{ url('/admin/events/create') }}" class="btn btn-success btn-sm" title="Add New Event">--}}
-{{--                        <i class="fa fa-plus" aria-hidden="true"></i> Add New--}}
-{{--                    </a>--}}
-{{--                </div>--}}
+                {{--                <div class="card-body">--}}
+                {{--                    <a href="{{ url('/admin/events/create') }}" class="btn btn-success btn-sm" title="Add New Event">--}}
+                {{--                        <i class="fa fa-plus" aria-hidden="true"></i> Add New--}}
+                {{--                    </a>--}}
+                {{--                </div>--}}
 
                 <br/>
                 <br/>
@@ -25,6 +25,7 @@
                                     <th>{{ucfirst($rule)}}</th>
                                 <?php endforeach; ?>
                                 <th>Action</th>
+                                <th>is_expired</th>
                             </tr>
                         </thead>
                     </table>
@@ -36,55 +37,58 @@
 </div>
 <script type="text/javascript">
     $(function () {
-    var table = $('.data-table').DataTable({
-    processing: true,
-            serverSide: true,
-            ajax: "{{ route('events.index') }}",
-            columns: [
-            {data: 'id', name: 'id'},
+        var table = $('.data-table').DataTable({
+        processing: true,
+                serverSide: true,
+                ajax: "{{ route('events.index') }}",
+                columns: [
+                {data: 'id', name: 'id'},
 <?php foreach ($rules as $rule): ?>
-                {data: "{{$rule}}", name: "{{$rule}}"},
+                    {data: "{{$rule}}", name: "{{$rule}}"},
 <?php endforeach; ?>
-            {data: 'action', name: 'action', orderable: false, searchable: false},
-            ]
+                {data: 'action', name: 'action', orderable: false, searchable: false},
+                {data: 'is_expired', name: 'is_expired', orderable: false, searchable: false}
+                ,
+
+                ]
     });
 //deleting data
-            $('.data-table').on('click', '.btnDelete[data-remove]', function (e) {
-    e.preventDefault();
-            var url = $(this).data('remove');
-            swal.fire({
+    $('.data-table').on('click', '.btnDelete[data-remove]', function (e) {
+        e.preventDefault();
+        var url = $(this).data('remove');
+        swal.fire({
             title: "Are you sure want to remove this item?",
-                    text: "Data will be Temporary Deleted!",
-                    type: "warning",
-                    showCancelButton: true,
-                    confirmButtonClass: "btn-danger",
-                    confirmButtonText: "Confirm",
-                    cancelButtonText: "Cancel",
-            }).then((result) => {
-    Swal.showLoading();
+            text: "Data will be Temporary Deleted!",
+            type: "warning",
+            showCancelButton: true,
+            confirmButtonClass: "btn-danger",
+            confirmButtonText: "Confirm",
+            cancelButtonText: "Cancel",
+        }).then((result) => {
+            Swal.showLoading();
             if (result.value) {
-    $.ajax({
-    url: url,
-            type: 'DELETE',
-            dataType: 'json',
-            data: {method: '_DELETE', submit: true, _token: '{{csrf_token()}}'},
-            success: function (data) {
-            if (data == 'Success') {
-            swal.fire("Deleted!", "Event has been deleted", "success");
-                    table.ajax.reload(null, false);
+                $.ajax({
+                    url: url,
+                    type: 'DELETE',
+                    dataType: 'json',
+                    data: {method: '_DELETE', submit: true, _token: '{{csrf_token()}}'},
+                    success: function (data) {
+                        if (data == 'Success') {
+                            swal.fire("Deleted!", "Event has been deleted", "success");
+                            table.ajax.reload(null, false);
+                        }
+                    }
+                });
             }
-            }
+        });
     });
-    }
-    });
-    });
-            $('.data-table').on('click', '.changeStatus', function (e) {
+    $('.data-table').on('click', '.changeStatus', function (e) {
     e.preventDefault();
             var id = $(this).attr('data-id');
             var status = $(this).attr('data-status');
             Swal.fire({
             title: 'Are you sure you wanted to change status?',
-                text: "You can revert this,in case you change your mind!",
+                    text: "You can revert this,in case you change your mind!",
                     type: 'warning',
                     showCancelButton: true,
                     confirmButtonColor: '#3085d6',
@@ -121,6 +125,7 @@
     }
     });
     });
-            });
+    }
+    );
 </script>
 @endsection
